Add Classpert under the MOOcs list
Recently, we launched https://classpert.com - a free search engine for online courses. Our main focus is on Computer Science and Data Science courses (we have indexed over 150k courses so far). I'm adding it under the MOOCs list, as I believe it might be a valuable resource for programmers.